Add files via upload
A handwritten digit recognition project using the MNIST dataset is a classic example of computer vision and machine learning. This project involves training a machine learning model, often a convolutional neural network (CNN), using the MNIST dataset, which consists of thousands of handwritten digits ranging from 0 to 9. The goal is to create a model that can accurately identify and classify these digits when presented with new, handwritten examples. This technology has a wide range of practical applications, from automated address reading in mail sorting to recognizing handwritten numbers on checks, making it a fundamental step in optical character recognition (OCR) and digit-based image analysis